standardized location name
