+++ booking/templates/booking/book_listing.html
@@ -18,7 +18,7 @@
         <div class="col-md-6">
           <p class="d-flex align-items-start mb-2">
             <i class="fas fa-map-marker-alt text-secondary me-2 mt-1"></i>
-            <span>{{ listing.location }}</span>
+            <span>{{ listing.location_name }}</span>
           </p>
           <p class="d-flex align-items-center mb-3">
             <i class="fas fa-tag text-secondary me-2"></i>
