+++ booking/templates/booking/my_bookings.html
@@ -34,7 +34,7 @@
                   <!-- Location -->
                   <p class="d-flex align-items-start mb-2">
                     <i class="fas fa-map-marker-alt text-secondary me-2 mt-1"></i>
-                    <span>{{ booking.listing.location }}</span>
+                    <span>{{ booking.listing.location_name }}</span>
                   </p>
                   
                   <!-- Price -->
