+++ listings/models.py
@@ -1,7 +1,10 @@
 import datetime as dt
-from django.db import models
+
 from django.contrib.auth.models import User
+from django.db import models
 from django.db.models import Max, Min
+
+from .utilities import simplify_location
 
 
 class Listing(models.Model):
@@ -13,13 +16,22 @@ class Listing(models.Model):
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
 
-    def average_rating(self):
+    @property
+    def location_name(self):
+        """Returns a simplified version of the location string."""
+        return simplify_location(self.location)
+
+    @property
+    def avg_rating(self):
+        """Returns the average rating for this listing."""
         reviews = self.reviews.all()
         if reviews:
             return sum(review.rating for review in reviews) / reviews.count()
         return None
 
+    @property
     def rating_count(self):
+        """Returns the total number of reviews for this listing."""
         return self.reviews.count()
 
     def __str__(self):
