+++ listings/templates/listings/manage_listings.html
@@ -40,7 +40,7 @@
                     <div class="mb-3">
                         <p class="d-flex align-items-center mb-2">
                             <i class="fas fa-map-marker-alt text-secondary me-2"></i>
-                            <span>{{ listing.location }}</span>
+                            <span>{{ listing.location_name }}</span>
                         </p>
                         
                         <p class="d-flex align-items-center mb-2">
